+++ RELEASE_NOTES.md
@@ -4,9 +4,10 @@ This file contains one line summaries of commits that are worthy of mentioning i
 A new header is inserted each time a *tag* is created.
 
 ## main branch
- 
-- engine: `Camera::getNear()` and `Camera::getCullingFar()` now return `doubles`
 
+## v1.25.2
+
+- engine: `Camera::getNear()` and `Camera::getCullingFar()` now return `doubles`
 - Metal: implement scissor support.
 - engine: `Renderer::getUserTime()` now returns seconds as documented (#5722) [⚠️ **API Fix**]